reduce number of noties
